Removed Redundant Code
