+++ src/lib.rs
@@ -4,12 +4,13 @@ use std::time::Duration;
 use thiserror::Error;
 
 pub mod anime;
+pub mod character;
 pub mod common;
 pub mod manga;
+pub mod misc;
+pub mod people;
 pub mod random;
 pub mod users;
-pub mod character;
-pub mod people;
 
 const API_BASE_URL: &str = "https://api.jikan.moe/v4";
