+++ src/people.rs
@@ -1,5 +1,5 @@
 // people.rs
-use crate::common::Pagination;
+use crate::common::{Images, Pagination};
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -8,5 +8,12 @@ pub struct PeopleResponse<T> {
     pub pagination: Option<Pagination>,
 }
 
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Person {
+    pub mal_id: i32,
+    pub url: String,
+    pub images: Images,
+    pub name: String,
+}
 
 // to implement
